Rename issue-unmap-dropdown-item component
