+++ src/ggrc-client/js/templates/audits/dropdown_menu.stache
@@ -77,11 +77,11 @@
   </li>
 
   {{#is_info_pin}}
-    <issue-unmap-dropdown-item
+    <unmap-dropdown-item
       {page_instance}="page_instance"
       {instance}="instance"
       {options}="options">
-    </issue-unmap-dropdown-item>
+    </unmap-dropdown-item>
 
     {{#if instance.viewLink}}
       <li>
